Share the SPI ARC-command adapter between the SPI devices
tt-metal compiles UMD as a unity build, which merges the two files'
anonymous namespaces into one translation unit, so the per-file copies of
the adapter collided (caught by the run-ttsim-tests job). One inline helper
in a shared internal header instead.

(cherry picked from commit a9419bf6f4221ffd476c7d582aeb28989d79d5fe)
