+++ device/arc/wormhole_spi_tt_device.cpp
@@ -15,6 +15,7 @@
 #include <string>
 #include <vector>
 
+#include "spi_arc_command.hpp"
 #include "umd/device/arc/arc_messenger.hpp"
 #include "umd/device/arc/arc_telemetry_reader.hpp"
 #include "umd/device/arch/wormhole_implementation.hpp"
@@ -26,25 +27,6 @@
 #include "umd/device/utils/semver.hpp"
 
 namespace tt::umd {
-
-namespace {
-
-// The SPI flows were written against ArcMessenger::send_message, which fills the caller's
-// return-value vector in place and returns the exit code. Kept as a local adapter so the conversion
-// to DeviceFirmware does not also rewrite the flows themselves.
-uint32_t send_arc_command(
-    DeviceFirmware* firmware,
-    uint32_t msg_code,
-    std::vector<uint32_t>& return_values,
-    const std::vector<uint32_t>& args = {}) {
-    DeviceCommandResult result = firmware->send_device_command(msg_code, args, timeout::ARC_MESSAGE_TIMEOUT);
-    for (size_t i = 0; i < return_values.size() && i < result.return_values.size(); i++) {
-        return_values[i] = result.return_values[i];
-    }
-    return result.exit_code;
-}
-
-}  // namespace
 
 // SPI Register Addresses (from Rust code).
 static constexpr uint64_t GPIO2_PAD_TRIEN_CNTL = wormhole::ARC_RESET_UNIT_OFFSET + 0x240;
@@ -326,7 +308,7 @@ void WormholeSPITTDevice::read(uint32_t addr, uint8_t* data, size_t size) {
     }
 
     std::vector<uint32_t> ret(1);
-    uint32_t rc = send_arc_command(
+    uint32_t rc = send_spi_arc_command(
         firmware,
         wormhole::ARC_MSG_COMMON_PREFIX | static_cast<uint32_t>(wormhole::arc_message_type::GET_SPI_DUMP_ADDR),
         ret);
@@ -351,7 +333,7 @@ void WormholeSPITTDevice::read(uint32_t addr, uint8_t* data, size_t size) {
 
         uint32_t spi_read_msg =
             wormhole::ARC_MSG_COMMON_PREFIX | static_cast<uint32_t>(wormhole::arc_message_type::SPI_READ);
-        send_arc_command(firmware, spi_read_msg, ret, {chunk_addr & 0xFFFF, (chunk_addr >> 16) & 0xFFFF});
+        send_spi_arc_command(firmware, spi_read_msg, ret, {chunk_addr & 0xFFFF, (chunk_addr >> 16) & 0xFFFF});
         device_->read_from_device(
             chunk_buf.data(),
             device_->get_arc_core(),
@@ -398,7 +380,7 @@ void WormholeSPITTDevice::write(uint32_t addr, const uint8_t* data, size_t size,
     std::exception_ptr write_exception;
     try {
         std::vector<uint32_t> ret(1);
-        uint32_t rc = send_arc_command(
+        uint32_t rc = send_spi_arc_command(
             firmware,
             wormhole::ARC_MSG_COMMON_PREFIX | static_cast<uint32_t>(wormhole::arc_message_type::GET_SPI_DUMP_ADDR),
             ret);
@@ -424,7 +406,7 @@ void WormholeSPITTDevice::write(uint32_t addr, const uint8_t* data, size_t size,
             // Read the current chunk first.
             uint32_t spi_read_msg =
                 wormhole::ARC_MSG_COMMON_PREFIX | static_cast<uint32_t>(wormhole::arc_message_type::SPI_READ);
-            send_arc_command(firmware, spi_read_msg, ret, {chunk_addr & 0xFFFF, (chunk_addr >> 16) & 0xFFFF});
+            send_spi_arc_command(firmware, spi_read_msg, ret, {chunk_addr & 0xFFFF, (chunk_addr >> 16) & 0xFFFF});
 
             device_->read_from_device(
                 chunk_buf.data(),
@@ -461,7 +443,7 @@ void WormholeSPITTDevice::write(uint32_t addr, const uint8_t* data, size_t size,
                 if (!skip_write_to_spi) {
                     uint32_t spi_write_msg =
                         wormhole::ARC_MSG_COMMON_PREFIX | static_cast<uint32_t>(wormhole::arc_message_type::SPI_WRITE);
-                    send_arc_command(firmware, spi_write_msg, ret, {0xFFFF, 0xFFFF});
+                    send_spi_arc_command(firmware, spi_write_msg, ret, {0xFFFF, 0xFFFF});
                 }
             }
         }
